refactor: change scope of room_model and room_solution to module; move geometric_fba_model to respective module
