+++ cobra/test/test_flux_analysis/test_geometric.py
@@ -7,7 +7,53 @@ from __future__ import absolute_import
 import numpy as np
 from pandas import Series
 
+import pytest
+from cobra.core import Metabolite, Model, Reaction, Solution
 from cobra.flux_analysis import geometric_fba
+
+
+@pytest.fixture(scope="module")
+def geometric_fba_model():
+    """
+    Generate geometric FBA model as described in [1]_
+
+    References
+    ----------
+    .. [1] Smallbone, Kieran & Simeonidis, Vangelis. (2009).
+           Flux balance analysis: A geometric perspective.
+           Journal of theoretical biology.258. 311-5.
+           10.1016/j.jtbi.2009.01.027.
+
+    """
+    test_model = Model('geometric_fba_paper_model')
+
+    test_model.add_metabolites(Metabolite('A'))
+    test_model.add_metabolites(Metabolite('B'))
+
+    v_1 = Reaction('v1', upper_bound=1.0)
+    v_1.add_metabolites({test_model.metabolites.A: 1.0})
+
+    v_2 = Reaction('v2', lower_bound=-1000.0)
+    v_2.add_metabolites({test_model.metabolites.A: -1.0,
+                         test_model.metabolites.B: 1.0})
+
+    v_3 = Reaction('v3', lower_bound=-1000.0)
+    v_3.add_metabolites({test_model.metabolites.A: -1.0,
+                         test_model.metabolites.B: 1.0})
+
+    v_4 = Reaction('v4', lower_bound=-1000.0)
+    v_4.add_metabolites({test_model.metabolites.A: -1.0,
+                         test_model.metabolites.B: 1.0})
+
+    v_5 = Reaction('v5')
+    v_5.add_metabolites({test_model.metabolites.A: 0.0,
+                         test_model.metabolites.B: -1.0})
+
+    test_model.add_reactions([v_1, v_2, v_3, v_4, v_5])
+
+    test_model.objective = 'v5'
+
+    return test_model
 
 
 def test_geometric_fba_benchmark(model, benchmark, all_solvers):
